Migrate grids over to EntryGrid
Allows us to re-use common UI across
Trending, Recommended and Popular
